Add legal guidance (#15)
* Updates robotframework-faker to version 5.0.0

This commit updates the robotframework-faker dependency to the latest version, 5.0.0.

This upgrade likely includes bug fixes, new features, and performance improvements, ensuring compatibility with other libraries and enhancing the functionality of the robot framework.

Enables shell script debugging

Turns on shell script debugging to aid in troubleshooting.

The `set -x` command is uncommented, which will print each command before it is executed.

Updates Python version and fixes upload script

Updates the Python version used in the CI/CD workflow to 3.13.3 for compatibility and to leverage the latest features.

Also fixes a bug in the device farm upload script where the script would not exit properly if the file upload failed, leading to incorrect CI/CD results. The fix ensures that the script exits with an error code and prints relevant debug data when an upload fails.

Sets platform for pip wheel building

Ensures that the pip wheel command builds wheels compatible
with the target platform in the CD workflow. This avoids
potential issues with platform-specific dependencies.

fix

* Sets pip to only use binary packages

Ensures that pip only uses binary packages during the deployment process.
This avoids potential issues with building packages from source and improves deployment reliability.

* fix: wheel dependencies

* dsaaa

* aaaaa

* se agrega el tests case de orientacion legal

* se quita una linea que se agrego por error

* se elimina linea que se agrego por error

---------

Co-authored-by: Christian <lopezchr@gmail.com>